+++ Dalamud/Interface/ImGuiNotification/INotification.cs
@@ -1,9 +1,10 @@
 using Dalamud.Interface.Internal.Notifications;
+using Dalamud.Plugin.Services;
 
 namespace Dalamud.Interface.ImGuiNotification;
 
 /// <summary>Represents a notification.</summary>
-public interface INotification : IDisposable
+public interface INotification
 {
     /// <summary>Gets or sets the content body of the notification.</summary>
     string Content { get; set; }
@@ -18,22 +19,13 @@ public interface INotification : IDisposable
     NotificationType Type { get; set; }
 
     /// <summary>Gets or sets the icon source.</summary>
-    /// <remarks>
-    /// <para>Assigning a new value that does not equal to the previous value will dispose the old value. The ownership
-    /// of the new value is transferred to this <see cref="INotification"/>. <b>Even if the assignment throws an
-    /// exception</b>, the ownership is transferred, causing the value to be disposed. Assignment should not throw an
-    /// exception though, so wrapping the assignment in try...catch block is not required.</para>
-    /// <para>The assigned value will be disposed upon the <see cref="IDisposable.Dispose"/> call on this instance of
-    /// <see cref="INotification"/>, unless the same value is assigned, in which case it will do nothing.</para>
-    /// <para>If this <see cref="INotification"/> is an <see cref="IActiveNotification"/>, then updating this property
-    /// will change the icon being displayed (calls <see cref="IActiveNotification.UpdateIcon"/>), unless
-    /// <see cref="IActiveNotification.IsDismissed"/> is <c>true</c>.</para>
-    /// </remarks>
-    INotificationIconSource? IconSource { get; set; }
+    /// <remarks>Use <see cref="IActiveNotification.SetIconTexture"/> to use a texture, after calling
+    /// <see cref="INotificationManager.AddNotification"/>.</remarks>
+    INotificationIcon? Icon { get; set; }
 
     /// <summary>Gets or sets the hard expiry.</summary>
     /// <remarks>
-    /// Setting this value will override <see cref="InitialDuration"/> and <see cref="DurationSinceLastInterest"/>, in that
+    /// Setting this value will override <see cref="InitialDuration"/> and <see cref="ExtensionDurationSinceLastInterest"/>, in that
     /// the notification will be dismissed when this expiry expires.<br />
     /// Set to <see cref="DateTime.MaxValue"/> to make only <see cref="InitialDuration"/> take effect.<br />
     /// If neither <see cref="HardExpiry"/> nor <see cref="InitialDuration"/> is not MaxValue, then the notification
@@ -45,7 +37,8 @@ public interface INotification : IDisposable
 
     /// <summary>Gets or sets the initial duration.</summary>
     /// <remarks>Set to <see cref="TimeSpan.MaxValue"/> to make only <see cref="HardExpiry"/> take effect.</remarks>
-    /// <remarks>Updating this value will reset the dismiss timer.</remarks>
+    /// <remarks>Updating this value will reset the dismiss timer, but the remaining duration will still be calculated
+    /// based on <see cref="IActiveNotification.CreatedAt"/>.</remarks>
     TimeSpan InitialDuration { get; set; }
 
     /// <summary>Gets or sets the new duration for this notification once the mouse cursor leaves the window and the
@@ -55,7 +48,7 @@ public interface INotification : IDisposable
     /// notification or focusing on it will not make the notification stay.<br />
     /// Updating this value will reset the dismiss timer.
     /// </remarks>
-    TimeSpan DurationSinceLastInterest { get; set; }
+    TimeSpan ExtensionDurationSinceLastInterest { get; set; }
 
     /// <summary>Gets or sets a value indicating whether to show an indeterminate expiration animation if
     /// <see cref="HardExpiry"/> is set to <see cref="DateTime.MaxValue"/>.</summary>
