+++ Dalamud/Interface/ImGuiNotification/NotificationUtilities.cs
@@ -17,44 +17,47 @@ namespace Dalamud.Interface.ImGuiNotification;
 /// <summary>Utilities for implementing stuff under <see cref="ImGuiNotification"/>.</summary>
 public static class NotificationUtilities
 {
-    /// <inheritdoc cref="INotificationIconSource.From(SeIconChar)"/>
+    /// <inheritdoc cref="INotificationIcon.From(SeIconChar)"/>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static INotificationIconSource ToIconSource(this SeIconChar iconChar) =>
-        INotificationIconSource.From(iconChar);
+    public static INotificationIcon ToIconSource(this SeIconChar iconChar) =>
+        INotificationIcon.From(iconChar);
 
-    /// <inheritdoc cref="INotificationIconSource.From(FontAwesomeIcon)"/>
+    /// <inheritdoc cref="INotificationIcon.From(FontAwesomeIcon)"/>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static INotificationIconSource ToIconSource(this FontAwesomeIcon iconChar) =>
-        INotificationIconSource.From(iconChar);
+    public static INotificationIcon ToIconSource(this FontAwesomeIcon iconChar) =>
+        INotificationIcon.From(iconChar);
 
-    /// <inheritdoc cref="INotificationIconSource.From(IDalamudTextureWrap,bool)"/>
+    /// <inheritdoc cref="INotificationIcon.FromFile(string)"/>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static INotificationIconSource ToIconSource(this IDalamudTextureWrap? wrap, bool takeOwnership = true) =>
-        INotificationIconSource.From(wrap, takeOwnership);
+    public static INotificationIcon ToIconSource(this FileInfo fileInfo) =>
+        INotificationIcon.FromFile(fileInfo.FullName);
 
-    /// <inheritdoc cref="INotificationIconSource.FromFile(string)"/>
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static INotificationIconSource ToIconSource(this FileInfo fileInfo) =>
-        INotificationIconSource.FromFile(fileInfo.FullName);
-
-    /// <summary>Draws an icon string.</summary>
-    /// <param name="fontHandleLarge">The font handle to use.</param>
-    /// <param name="c">The icon character.</param>
+    /// <summary>Draws an icon from an <see cref="IFontHandle"/> and a <see cref="char"/>.</summary>
     /// <param name="minCoord">The coordinates of the top left of the icon area.</param>
     /// <param name="maxCoord">The coordinates of the bottom right of the icon area.</param>
+    /// <param name="c">The icon character.</param>
+    /// <param name="fontHandle">The font handle to use.</param>
     /// <param name="color">The foreground color.</param>
-    internal static unsafe void DrawIconString(
-        IFontHandle fontHandleLarge,
-        char c,
+    /// <returns><c>true</c> if anything has been drawn.</returns>
+    internal static unsafe bool DrawIconFrom(
         Vector2 minCoord,
         Vector2 maxCoord,
+        char c,
+        IFontHandle fontHandle,
         Vector4 color)
     {
+        if (c is '\0' or char.MaxValue)
+            return false;
+
         var smallerDim = Math.Max(maxCoord.Y - minCoord.Y, maxCoord.X - minCoord.X);
-        using (fontHandleLarge.Push())
+        using (fontHandle.Push())
         {
             var font = ImGui.GetFont();
-            ref readonly var glyph = ref *(ImGuiHelpers.ImFontGlyphReal*)font.FindGlyph(c).NativePtr;
+            var glyphPtr = (ImGuiHelpers.ImFontGlyphReal*)font.FindGlyphNoFallback(c).NativePtr;
+            if (glyphPtr is null)
+                return false;
+
+            ref readonly var glyph = ref *glyphPtr;
             var size = glyph.XY1 - glyph.XY0;
             var smallerSizeDim = Math.Min(size.X, size.Y);
             var scale = smallerSizeDim > smallerDim ? smallerDim / smallerSizeDim : 1f;
@@ -69,67 +72,72 @@ public static class NotificationUtilities
                 glyph.UV1,
                 ImGui.GetColorU32(color with { W = color.W * ImGui.GetStyle().Alpha }));
         }
+
+        return true;
     }
 
-    /// <summary>Draws the given texture, or the icon of the plugin if texture is <c>null</c>.</summary>
-    /// <param name="texture">The texture.</param>
+    /// <summary>Draws an icon from an instance of <see cref="IDalamudTextureWrap"/>.</summary>
     /// <param name="minCoord">The coordinates of the top left of the icon area.</param>
     /// <param name="maxCoord">The coordinates of the bottom right of the icon area.</param>
-    /// <param name="initiatorPlugin">The initiator plugin.</param>
-    internal static void DrawTexture(
-        IDalamudTextureWrap? texture,
-        Vector2 minCoord,
-        Vector2 maxCoord,
-        LocalPlugin? initiatorPlugin)
+    /// <param name="texture">The texture.</param>
+    /// <returns><c>true</c> if anything has been drawn.</returns>
+    internal static bool DrawIconFrom(Vector2 minCoord, Vector2 maxCoord, IDalamudTextureWrap? texture)
     {
-        var handle = nint.Zero;
-        var size = Vector2.Zero;
-        if (texture is not null)
+        if (texture is null)
+            return false;
+        try
         {
-            try
+            var handle = texture.ImGuiHandle;
+            var size = texture.Size;
+            if (size.X > maxCoord.X - minCoord.X)
+                size *= (maxCoord.X - minCoord.X) / size.X;
+            if (size.Y > maxCoord.Y - minCoord.Y)
+                size *= (maxCoord.Y - minCoord.Y) / size.Y;
+            ImGui.SetCursorPos(((minCoord + maxCoord) - size) / 2);
+            ImGui.Image(handle, size);
+            return true;
+        }
+        catch
+        {
+            return false;
+        }
+    }
+
+    /// <summary>Draws an icon from an instance of <see cref="LocalPlugin"/>.</summary>
+    /// <param name="minCoord">The coordinates of the top left of the icon area.</param>
+    /// <param name="maxCoord">The coordinates of the bottom right of the icon area.</param>
+    /// <param name="plugin">The plugin. Dalamud icon will be drawn if <c>null</c> is given.</param>
+    /// <returns><c>true</c> if anything has been drawn.</returns>
+    internal static bool DrawIconFrom(Vector2 minCoord, Vector2 maxCoord, LocalPlugin? plugin)
+    {
+        var dam = Service<DalamudAssetManager>.Get();
+        if (plugin is null)
+            return false;
+
+        if (!Service<PluginImageCache>.Get().TryGetIcon(
+                plugin,
+                plugin.Manifest,
+                plugin.IsThirdParty,
+                out var texture) || texture is null)
+        {
+            texture = plugin switch
             {
-                handle = texture.ImGuiHandle;
-                size = texture.Size;
-            }
-            catch
-            {
-                // must have been disposed or something; ignore the texture
-            }
+                { IsDev: true } => dam.GetDalamudTextureWrap(DalamudAsset.DevPluginIcon),
+                { IsThirdParty: true } => dam.GetDalamudTextureWrap(DalamudAsset.ThirdInstalledIcon),
+                _ => dam.GetDalamudTextureWrap(DalamudAsset.InstalledIcon),
+            };
         }
 
-        if (handle == nint.Zero)
-        {
-            var dam = Service<DalamudAssetManager>.Get();
-            if (initiatorPlugin is null)
-            {
-                texture = dam.GetDalamudTextureWrap(DalamudAsset.LogoSmall);
-            }
-            else
-            {
-                if (!Service<PluginImageCache>.Get().TryGetIcon(
-                        initiatorPlugin,
-                        initiatorPlugin.Manifest,
-                        initiatorPlugin.IsThirdParty,
-                        out texture) || texture is null)
-                {
-                    texture = initiatorPlugin switch
-                    {
-                        { IsDev: true } => dam.GetDalamudTextureWrap(DalamudAsset.DevPluginIcon),
-                        { IsThirdParty: true } => dam.GetDalamudTextureWrap(DalamudAsset.ThirdInstalledIcon),
-                        _ => dam.GetDalamudTextureWrap(DalamudAsset.InstalledIcon),
-                    };
-                }
-            }
+        return DrawIconFrom(minCoord, maxCoord, texture);
+    }
 
-            handle = texture.ImGuiHandle;
-            size = texture.Size;
-        }
-
-        if (size.X > maxCoord.X - minCoord.X)
-            size *= (maxCoord.X - minCoord.X) / size.X;
-        if (size.Y > maxCoord.Y - minCoord.Y)
-            size *= (maxCoord.Y - minCoord.Y) / size.Y;
-        ImGui.SetCursorPos(((minCoord + maxCoord) - size) / 2);
-        ImGui.Image(handle, size);
+    /// <summary>Draws the Dalamud logo as an icon.</summary>
+    /// <param name="minCoord">The coordinates of the top left of the icon area.</param>
+    /// <param name="maxCoord">The coordinates of the bottom right of the icon area.</param>
+    internal static void DrawIconFromDalamudLogo(Vector2 minCoord, Vector2 maxCoord)
+    {
+        var dam = Service<DalamudAssetManager>.Get();
+        var texture = dam.GetDalamudTextureWrap(DalamudAsset.LogoSmall);
+        DrawIconFrom(minCoord, maxCoord, texture);
     }
 }
