+++ Dalamud/Interface/UiBuilder.cs
@@ -581,7 +581,6 @@ public sealed class UiBuilder : IDisposable
                 Type = type,
                 InitialDuration = TimeSpan.FromMilliseconds(msDelay),
             },
-            true,
             this.localPlugin);
         _ = this.notifications.TryAdd(an, 0);
         an.Dismiss += (a, unused) => this.notifications.TryRemove(an, out _);
